Refactor backend prompt engineering code
