+++ digiprod_gen/backend_api/prompt_engineering/midjourney_prompt_examples.py
